% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\docType{package}
\encoding{UTF-8}
\name{hassediagram-package}
\alias{hassediagram-package}
\title{hassediagram: Hasse diagram of the layout structure and restricted layout structure}
\description{
Returns a Hasse diagram of the 'layout structure' or the 'restricted layout structure' of an experimental design.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GSK-Biostatistics/hassediagram}
  \item Report bugs at \url{https://github.com/GSK-Biostatistics/hassediagram/issues}
}

}
\author{
\strong{Maintainer}: Damianos Michaelides \email{dm3g15@soton.ac.uk}

Authors:
\itemize{
  \item Simon Bate \email{simon.t.bate@gsk.com}
  \item Marion Chatfield
}

}
\keyword{internal}
