% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\docType{package}
\encoding{UTF-8}
\name{hassediagrams-package}
\alias{hassediagrams-package}
\title{hassediagrams: Hasse Diagram of the Layout Structure and Restricted Layout Structure}
\description{
Returns a Hasse diagram of the layout structure (Bate and Chatfield (2016)) \doi{10.1080/00224065.2016.11918173} or the restricted layout structure (Bate and Chatfield (2016)) \doi{10.1080/00224065.2016.11918174} of an experimental design.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GSK-Biostatistics/hassediagrams}
  \item Report bugs at \url{https://github.com/GSK-Biostatistics/hassediagrams/issues}
}

}
\author{
\strong{Maintainer}: Damianos Michaelides \email{dm3g15@soton.ac.uk}

Authors:
\itemize{
  \item Simon Bate \email{simon.t.bate@gsk.com}
  \item Marion Chatfield
}

}
\keyword{internal}
