% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary_rls.R
\name{print.rls}
\alias{print.rls}
\alias{summary.rls}
\title{Print and Summary of \code{rls} objects}
\usage{
\method{print}{rls}(x, ...)

\method{summary}{rls}(object, ...)
}
\arguments{
\item{x}{An object of class \code{"rls"}.}

\item{...}{Additional arguments to be passed to other methods.}

\item{object}{An object of class \code{"rls"}.}
}
\value{
The functions return the structural objects of the Layout Structure 
and a matrix that the users need to edit with the randomisation 
objects to pass in the \code{\link[hassediagrams]{hasserls}} function.
}
\description{
Print and Summary of objects of class \code{"rls"}.
}
\note{
For examples see \code{\link[hassediagrams]{itemlist}} and \code{\link[hassediagrams]{hasserls}}.
}
\author{
Damianos Michaelides, Simon Bate, and Marion Chatfield
}
